\name{penalizedLDA-internal}
\title{Internal penalizedLDA functions}
\alias{print.penlda}
\alias{plot.penlda}
\alias{print.penldacv}
\alias{plot.penldacv}
\description{Internal penalizedLDA functions}
\usage{
plot.penlda(x,...)
print.penlda(x,...)
plot.penldacv(x,...)
print.penldacv(x,...)
}
\author{Daniela M. Witten}
\keyword{internal}



